Update DDF build using Template Pattern, Add BigQuery DataSource & BigQueryDDF

+++ NewDDF/ddf-core/src/main/java/io/ddf2/SqlResult.java
@@ -149,7 +149,7 @@ public class SqlResult implements ISqlResult {
     }
 
     @Override
-    public long getDouble(int index, long defaultValue) {
+    public double getDouble(int index, double defaultValue) {
         return 0;
     }
